Hazards technique
Thanks to Chris for the work.

I entered the technique of hazards. In order to handle the different permutations of possible hazards, I thought of using an array. I created a "Join array to comma list" function to concatenate the array (thanks to Matt for support).

This is a starting point. As an editor we still have some doubts, which we will dissolve in the call.