Delete the definition of known to have accessibility limitations
There is no common way to express that a standard is not reached (i.e.: I cannot say "not reaching WCAG A").
The accessibility limitations will already be informed for the two key reading mode information (non visual and display transformability).
The way to express accessibility limitations can be discussed in a separate issue.