Adds Additional accessibility information technics IDs as Key to the JSON
The delta between Principle examples and Technics for this section is making alignement impossible.

Probable best option is to rewrite this example section of the principles. See #392

NOte that I maintained previews Key values coresponding to the actual examples in the principles. To update once remediation is done and Principle examples are informed as technics.